+++ examples/gis_import.py
@@ -2,4 +2,3 @@ from swmm_api.input_file.macros.gis import gpkg_to_swmm
 
 if __name__ == '__main__':
     inp = gpkg_to_swmm('/mnt/Windows/Users/mp/GIS/inp_update.gpkg')
-    print()