+++ examples/inp_file_reader.py
@@ -5,6 +5,9 @@ from swmm_api.input_file.sections.map_geodata import (VerticesGeo, CoordinateGeo
 from swmm_api.input_file.macros.geo import update_vertices
 from swmm_api.run import swmm5_run
 
+inp = read_inp_file('/home/markus/PycharmProjects/swmm_api/examples/internal/2015_06_17_UG_Weiz_OPTI_maxAbk_Ret_3J_60_KW.inp')
+exit()
+
 inp = read_inp_file('epaswmm5_apps_manual/Example7-Final.inp',
                     custom_converter={VERTICES: VerticesGeo,
                                       COORDINATES: CoordinateGeo,